build: apply upstream cmark change for python support
Original commit message:
~~~
build: add an upgrade path for newer CMake

CMake 3.12 deprecated `FindPythonInterp`, and with CMake 3.27, were
obsoleted with CMP0148.  Add a version check and switch to the new
behaviour to allow building with newer releases.  When the minimum CMake
version is increased, we can look at dropping the compatibility path.
~~~