remove `HAVE_UNISTD_H` in favour of compiler checks
Use the compiler provided macros to check if we are on a Unix system.
`unistd.h` is the Unix standard header which must be provided by any
Unix system. Apple macOS X is special in that it does not define
`__unix__` but is a Unix system. Sink this into the code rather than
keeping it in the build system to allow us to remove the generated
configuration headers.