glob: don't glob the headers
Globbing does not help the build system - it does not know the files.
The glob is only performed at configure time, anything done after that
is not reflected in the build system. Just simply ignore the header
changes as it is pointless. Let the header scan capture the dependencies
if needed.